pyproject.toml
[project]
name = "agent-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "attrs==25.4.0",
    "authlib==1.6.5",
    "backports-tarfile==1.2.0 ; python_full_version < '3.12'",
    "beartype==0.22.5",
    "boto3>=1.41.2",
    "cachetools==6.2.2",
    "certifi==2025.11.12",
    "cffi==2.0.0 ; platform_python_implementation != 'PyPy'",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6 ; sys_platform == 'win32'",
    "cryptography==46.0.3",
    "cyclopts==4.2.4",
    "diskcache==5.6.3",
    "dnspython==2.8.0",
    "docstring-parser==0.17.0",
    "docutils==0.22.3",
    "dotenv==0.9.9",
    "email-validator==2.3.0",
    "exceptiongroup==1.3.0",
    "fastapi==0.121.2",
    "fastmcp==2.13.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "idna==3.11",
    "importlib-metadata==8.7.0 ; python_full_version < '3.12'",
    "jaraco-classes==3.4.0",
    "jaraco-context==6.0.1",
    "jaraco-functools==4.3.0",
    "jeepney==0.9.0 ; sys_platform == 'linux'",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-path==0.3.4",
    "jsonschema-specifications==2025.9.1",
    "keyring==25.7.0",
    "langchain==1.0.7",
    "langchain-core==1.0.5",
    "langgraph==1.0.3",
    "langgraph-checkpoint==3.0.1",
    "langgraph-prebuilt==1.0.4",
    "langgraph-sdk==0.2.9",
    "langsmith==0.4.43",
    "markdown-it-py==4.0.0",
    "mcp==1.21.2",
    "mdurl==0.1.2",
    "more-itertools==10.8.0",
    "openapi-pydantic==0.5.1",
    "orjson==3.11.4",
    "ormsgpack==1.12.0",
    "packaging==25.0",
    "pathable==0.4.4",
    "pathvalidate==3.3.1",
    "platformdirs==4.5.0",
    "py-key-value-aio==0.2.8",
    "py-key-value-shared==0.2.8",
    "pycparser==2.23 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pyperclip==1.11.0",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.20",
    "pywin32==311 ; sys_platform == 'win32'",
    "pywin32-ctypes==0.2.3 ; sys_platform == 'win32'",
    "pyyaml==6.0.3",
    "referencing==0.36.2",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rich==14.2.0",
    "rich-rst==1.3.2",
    "rpds-py==0.29.0",
    "secretstorage==3.4.1 ; sys_platform == 'linux'",
    "sniffio==1.3.1",
    "sse-starlette==3.0.3",
    "starlette==0.49.3",
    "tenacity==9.1.2",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "websockets==15.0.1",
    "xxhash==3.6.0",
    "zipp==3.23.0 ; python_full_version < '3.12'",
    "zstandard==0.25.0",
]
